# backend/auth/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Это API лабораторной работы 1 по дисциплине \"Разработка бизнес-приложений\".\nТема проекта - аэропорт.",
        "title": "Разработка бизнес-приложений - лаба 1",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
...
        "/tickets/booking": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Book tickets",
                "parameters": [
                    {
                        "description": "example",
                        "name": "TicketsBookingParams",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TicketsBookingParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TicketsBookResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/tickets/confirm": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Confirm tickets",
                "parameters": [
                    {
                        "description": "example",
                        "name": "TicketIDsParams",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TicketIDsParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/tickets/remaining-seats-count": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Check enough tickets",
                "parameters": [
                    {
                        "description": "example",
                        "name": "TicketsCountRemainingSeatsParams",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TicketsCountRemainingSeatsParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TicketsRemainingSeatsCount"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Aircraft": {
            "type": "object",
            "properties": {
                "business_seats": {
                    "type": "integer"
                },
                "economy_seats": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "make_model": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "total_seats": {
                    "type": "integer"
                }
            }
        },
        "models.Airport": {
            "type": "object",
            "properties": {
                "country_id": {
                    "type": "integer"
                },
                "iata_code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.CabinType": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.Route": {
            "type": "object",
            "properties": {
                "arrival_airport": {
                    "$ref": "#/definitions/models.Airport"
                },
                "arrival_airport_id": {
                    "type": "integer"
                },
                "departure_airport": {
                    "$ref": "#/definitions/models.Airport"
                },
                "departure_airport_id": {
                    "type": "integer"
                },
                "distance": {
                    "type": "integer"
                },
                "flight_time": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "models.Schedule": {
            "type": "object",
            "properties": {
                "aircraft": {
                    "$ref": "#/definitions/models.Aircraft"
                },
                "aircraft_id": {
                    "type": "integer"
                },
                "confirmed": {
                    "type": "boolean"
                },
                "date": {
                    "type": "string"
                },
                "economy_price": {
                    "type": "number"
                },
                "flight_number": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "outbound": {
                    "type": "string"
                },
                "route": {
                    "$ref": "#/definitions/models.Route"
                },
                "route_id": {
                    "type": "integer"
                },
                "time": {
                    "type": "string"
                }
            }
        },
...
        "models.Ticket": {
            "type": "object",
            "properties": {
                "booking_reference": {
                    "type": "string"
                },
                "cabinType": {
                    "$ref": "#/definitions/models.CabinType"
                },
                "cabin_type_id": {
                    "type": "integer"
                },
                "confirmed": {
                    "type": "boolean"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "passport_country_id": {
                    "type": "integer"
                },
                "passport_number": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "schedule": {
                    "$ref": "#/definitions/models.Schedule"
                },
                "schedule_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.TicketIDsParams": {
            "type": "object",
            "properties": {
                "tickets": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "models.TicketPassengerInfo": {
            "type": "object",
            "properties": {
                "birthday": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "passport_country_id": {
                    "type": "integer"
                },
                "passport_number": {
                    "type": "string"
                },
                "phone": {
                    "description": "Email string `json:\"email\"`",
                    "type": "string"
                }
            }
        },
        "models.TicketsBookResult": {
            "type": "object",
            "properties": {
                "tickets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Ticket"
                    }
                },
                "total_cost": {
                    "type": "number"
                }
            }
        },
        "models.TicketsBookingParams": {
            "type": "object",
            "properties": {
                "cabin_type": {
                    "type": "integer"
                },
                "inbound_schedules": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "outbound_schedules": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "passengers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TicketPassengerInfo"
                    }
                }
            }
        },
        "models.TicketsCountRemainingSeatsParams": {
            "type": "object",
            "properties": {
                "schedule_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "models.TicketsRemainingSeatsCount": {
            "type": "object",
            "properties": {
                "business_seats": {
                    "type": "integer"
                },
                "economy_seats": {
                    "type": "integer"
                },
                "first_class_seats": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "access-token",
            "in": "cookie"
        }
    }
}